Fix some compiler warnings
git-svn-id: https://smokinguns.svn.sourceforge.net/svnroot/smokinguns/trunk@115 07eb42f0-b032-4355-afd3-42b1f3832223

+++ code/game/q_shared.c
@@ -887,7 +887,7 @@ void QDECL Com_sprintf( char *dest, int size, const char *fmt, ...) {
 	va_start (argptr,fmt);
 	len = vsprintf (bigbuffer,fmt,argptr);
 	va_end (argptr);
-	if ( len >= sizeof( bigbuffer ) ) {
+	if ( len >= (int)sizeof( bigbuffer ) ) {
 		Com_Error( ERR_FATAL, "Com_sprintf: overflowed bigbuffer" );
 	}
 	if (len >= size) {
